requirements.txt
apache-airflow-providers-google
lxml
beautifulsoup4
astronomer-providers
pandas
